refactor: modularize Objects base class with Mixins
Decomposed the monolithic Objects base class in apts/objects/base.py
into single-responsibility VisibilityMixIn (in apts/objects/visibility.py)
and AlmanacMixIn (in apts/objects/almanac.py) classes, keeping the
public API exactly consistent and highly maintainable.
